implemented retrieving food from dish

+++ src/main/java/com/example/demo/repositories/FoodRepository.java
@@ -3,5 +3,8 @@ package com.example.demo.repositories;
 import com.example.demo.entities.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodRepository extends JpaRepository<Food, Long> {
+import java.util.List;
+
+public interface FoodRepository extends JpaRepository<Food, Integer> {
+
 }
